(feat) add authorization

+++ services/droplet_service.py
@@ -1,5 +1,6 @@
 from fastapi import HTTPException
 import requests
+from models.user import UserInDB
 from utils.config import get_settings
 
 from utils.get_droplet_config import extract_database_info, get_payload_request
@@ -10,7 +11,12 @@ DIGITALOCEAN_API_URL = "https://api.digitalocean.com/v2/databases"
 
 
 class DropletService:
-    def create_droplet(username: str, database_type: str):
+    def create_droplet(user: UserInDB, database_type: str):
+        if len(user.droplets) >= 2:
+            raise HTTPException(
+                status_code=400,
+                detail="You have reached the maximum number of databases"
+            )
         if database_type is None:
             raise HTTPException(status_code=400, detail="Database type is required")
         database_type = database_type.lower()
@@ -21,7 +27,7 @@ class DropletService:
             )
         try:
             payload = get_payload_request(
-                username=username, database_type=database_type
+                username=user.username, database_type=database_type
             )
             response = requests.post(
                 DIGITALOCEAN_API_URL, headers=headers, json=payload
@@ -35,7 +41,11 @@ class DropletService:
         except Exception as e:
             raise HTTPException(status_code=response.status_code, detail=e.__dict__)
 
-    def get_droplet(droplet_id: str):
+    def get_droplet( user: UserInDB,droplet_id: str):
+        if droplet_id not in user.droplets:
+            raise HTTPException(
+                status_code=404, detail=f"Droplet with id {droplet_id} not found"
+            )
         try:
             response = requests.get(
                 f"{DIGITALOCEAN_API_URL}/{droplet_id}", headers=headers
@@ -49,7 +59,11 @@ class DropletService:
         except Exception as e:
             raise HTTPException(status_code=response.status_code, detail=e.__dict__)
 
-    def delete_droplet(droplet_id: str):
+    def delete_droplet(user: UserInDB, droplet_id: str):
+        if droplet_id not in user.droplets:
+            raise HTTPException(
+                status_code=404, detail=f"Droplet with id {droplet_id} not found"
+            )
         try:
             response = requests.delete(
                 f"{DIGITALOCEAN_API_URL}/{droplet_id}", headers=headers
@@ -62,3 +76,4 @@ class DropletService:
                 )
         except Exception as e:
             raise HTTPException(status_code=response.status_code, detail=e.__dict__)
+
